# BACKEND/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Loan Application API
  description: >
    Public API for submitting loan applications. There is no applicant account system,
    no login, and no retrieval endpoint — see applications-api.md for the security model.
  version: "1.0.0"
servers:
  - url: http://localhost:5001/api
    description: Local development (matches PORT in .env)
  - url: https://meridianloan.vercel.app/api
    description: Production (same origin as the frontend — served by a Vercel Function)

paths:
  /applications/uploads/init:
    post:
      operationId: initApplicationUploads
      summary: Issue an applicationId and signed direct-to-Cloudinary upload targets
      description: >
        Step 1 of submission. The three required verification images are uploaded directly
        from the browser to Cloudinary (never through this API — a Vercel Function request body
        is capped at 4.5MB), using the signed parameters this endpoint returns. Step 2 is the
        browser uploading to Cloudinary with them; step 3 is POST /applications below.
      responses:
        "201":
          description: Upload targets issued.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  applicationId: { type: string, example: LN-20260913-A7K4P9X }
                  uploads:
                    type: object
                    properties:
                      idCardImage: { $ref: '#/components/schemas/CloudinaryUploadTarget' }
                      ssnCardImage: { $ref: '#/components/schemas/CloudinaryUploadTarget' }
                      selfieImage: { $ref: '#/components/schemas/CloudinaryUploadTarget' }
                required: [success, applicationId, uploads]
        "429":
          description: Rate limit exceeded for this IP.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  /applications:
    get:
      operationId: listApplications_DOES_NOT_EXIST
      summary: "(Intentionally absent) — expect 404"
      description: >
        There is no public listing/enumeration endpoint. Nothing under this API returns more
        than one applicant's data. This entry exists only so you can confirm the 404 in EchoAPI.
      responses:
        "404":
          description: Not found, by design.
    post:
      operationId: createApplication
      summary: Submit a new loan application
      description: >
        Step 3 of submission (see POST /applications/uploads/init above) — plain JSON, no file
        bytes. applicationId must be one returned by that endpoint, with all three images
        already uploaded to Cloudinary under it; the server fetches them back from Cloudinary
        itself and rejects the request (VALIDATION_ERROR) if any is missing or fails to decode
        as an image. Any field not defined in the ApplicationSubmission schema (status,
        creditScore, adminNotes, riskScore, interestRate, approvalStatus, etc.), at any nesting
        level, is rejected with VALIDATION_ERROR rather than stored.
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          description: >
            8-128 chars, [A-Za-z0-9_-]. Send the same key when retrying a submission
            (double-click, network retry) to get back the original result instead of creating a
            second application.
          schema:
            type: string
            pattern: '^[A-Za-z0-9_-]{8,128}$'
            example: test-key-12345678
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApplicationSubmission' }
      responses:
        "201":
          description: Application submitted successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: Application submitted successfully. }
                  applicationId: { type: string, example: LN-20260913-A7K4P9X }
                required: [success, message, applicationId]
        "400":
          description: >
            Validation error in the "data" JSON, malformed "data" JSON, missing/invalid
            Idempotency-Key, a missing required image field, or an image that is not a
            valid/decodable image (INVALID_IMAGE).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "403":
          description: Origin not allowed by CORS policy.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "413":
          description: One of the uploaded images exceeds MAX_DOCUMENT_UPLOAD_SIZE_BYTES.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "429":
          description: Rate limit exceeded for this IP.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "500":
          description: Unexpected server error (no internal detail is ever included).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  /applications/{id}:
    get:
      operationId: getApplicationById_DOES_NOT_EXIST
      summary: "(Intentionally absent) — expect 404"
      description: >
        There is no public retrieval endpoint. A public applicationId must never by itself
        grant read access to applicant data (IDOR/BOLA prevention). This entry exists only so
        you can confirm the 404 in EchoAPI.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "404":
          description: Not found, by design.

  /applications/{applicationId}/documents/init:
    post:
      operationId: initDocumentUpload
      summary: Issue a signed direct-to-Cloudinary upload target for one document retake
      description: >
        Step 1 of a retake — same reasoning as POST /applications/uploads/init above, scoped to
        a single document kind on an application that already exists.
      parameters:
        - in: path
          name: applicationId
          required: true
          schema: { type: string, example: LN-20260913-A7K4P9X }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind]
              properties:
                kind: { type: string, enum: [id_card, ssn_card, selfie] }
      responses:
        "201":
          description: Upload target issued.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  upload: { $ref: '#/components/schemas/CloudinaryUploadTarget' }
        "400":
          description: Invalid kind.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "404":
          description: No application exists for this applicationId.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "429":
          description: Rate limit exceeded for this IP.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  /applications/{applicationId}/documents:
    post:
      operationId: replaceApplicationDocument
      summary: Confirm a replaced verification image on an existing application
      description: >
        Step 2 of a retake (see POST /applications/{applicationId}/documents/init above) — the
        image itself was already uploaded directly to Cloudinary; this call just confirms the
        kind so the server can fetch it back, process it, and update MongoDB. applicationId is
        the only access control: it grants no read access to any applicant data, only
        permission to replace one image on that specific application. Same IDOR caveat as the
        Idempotency-Key design: do not rely on this alone if the threat model changes.
      parameters:
        - in: path
          name: applicationId
          required: true
          schema: { type: string, example: LN-20260913-A7K4P9X }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind]
              properties:
                kind: { type: string, enum: [id_card, ssn_card, selfie] }
      responses:
        "201":
          description: Document replaced successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string }
                  kind: { type: string, enum: [id_card, ssn_card, selfie] }
        "400":
          description: Invalid kind, missing image, or image fails to decode (INVALID_IMAGE).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "404":
          description: No application exists for this applicationId.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "413":
          description: Image exceeds MAX_DOCUMENT_UPLOAD_SIZE_BYTES.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "429":
          description: Rate limit exceeded for this IP.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

components:
  schemas:
    ApplicationSubmission:
      type: object
      additionalProperties: false
      required: [applicationId, applicant, employment, loanRequest, disbursement, consent]
      properties:
        applicationId:
          type: string
          example: LN-20260913-A7K4P9X
          description: >
            Must be a value returned by POST /applications/uploads/init, with all three
            required images already uploaded to Cloudinary under it.
        applicant: { $ref: '#/components/schemas/Applicant' }
        employment: { $ref: '#/components/schemas/Employment' }
        loanRequest: { $ref: '#/components/schemas/LoanRequest' }
        loanHistory:
          type: array
          maxItems: 20
          default: []
          items: { $ref: '#/components/schemas/LoanHistoryEntry' }
        disbursement: { $ref: '#/components/schemas/Disbursement' }
        consent: { $ref: '#/components/schemas/Consent' }

    Applicant:
      type: object
      additionalProperties: false
      required: [firstName, lastName, email, phoneNumber, dateOfBirth, residentialAddress, city, state]
      properties:
        firstName: { type: string, maxLength: 80 }
        lastName: { type: string, maxLength: 80 }
        email: { type: string, format: email, maxLength: 254 }
        phoneNumber: { type: string, pattern: '^[0-9+\-() ]{7,20}$' }
        dateOfBirth: { type: string, format: date, description: "Applicant must be 18-120 years old." }
        gender: { type: string, enum: [male, female, other, prefer_not_to_say] }
        residentialAddress: { type: string, maxLength: 250 }
        city: { type: string, maxLength: 100 }
        state: { type: string, maxLength: 100 }

    Employment:
      type: object
      additionalProperties: false
      required: [employmentStatus]
      description: >
        employerName is required when employmentStatus is 'employed'. monthlyIncome and
        incomeFrequency are required when employmentStatus is 'employed' or 'self_employed'.
      properties:
        employmentStatus: { type: string, enum: [employed, self_employed, unemployed, student, retired] }
        employerName: { type: string, maxLength: 150 }
        jobTitle: { type: string, maxLength: 100 }
        employmentDurationMonths: { type: integer, minimum: 0, maximum: 1200 }
        monthlyIncome: { type: number, exclusiveMinimum: 0 }
        incomeFrequency: { type: string, enum: [weekly, biweekly, monthly, annually] }

    LoanRequest:
      type: object
      additionalProperties: false
      required: [requestedLoanAmount, loanPurpose, preferredRepaymentPeriodMonths, repaymentFrequency]
      properties:
        requestedLoanAmount: { type: number, exclusiveMinimum: 0 }
        loanPurpose: { type: string, maxLength: 200 }
        preferredRepaymentPeriodMonths: { type: integer, minimum: 1, maximum: 360 }
        repaymentFrequency: { type: string, enum: [weekly, biweekly, monthly, quarterly, annually] }

    LoanHistoryEntry:
      type: object
      additionalProperties: false
      required: [lenderName, originalLoanAmount, repaymentStatus, startDate]
      description: "endDate must be >= startDate. outstandingAmount must be <= originalLoanAmount."
      properties:
        lenderName: { type: string, maxLength: 150 }
        loanType: { type: string, maxLength: 100 }
        originalLoanAmount: { type: number, exclusiveMinimum: 0 }
        outstandingAmount: { type: number, minimum: 0 }
        repaymentStatus: { type: string, enum: [active, paid, overdue, defaulted, written_off] }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        repaymentFrequency: { type: string, enum: [weekly, biweekly, monthly, quarterly, annually] }
        monthlyPayment: { type: number, minimum: 0 }
        purpose: { type: string, maxLength: 200 }

    Disbursement:
      type: object
      additionalProperties: false
      required: [preferredMethod, bankDetails]
      description: >
        bankDetails is always required regardless of preferredMethod. otherMethodDetails is
        required when preferredMethod is 'other'.
      properties:
        preferredMethod: { type: string, enum: [check, direct_deposit, wire_transfer, ach, other] }
        otherMethodDetails: { type: string, maxLength: 200 }
        bankDetails: { $ref: '#/components/schemas/BankDetails' }

    BankDetails:
      type: object
      additionalProperties: false
      required: [accountHolderName, bankRoutingNumber, accountNumber, accountType, bankType]
      properties:
        accountHolderName: { type: string, maxLength: 150 }
        bankRoutingNumber:
          type: string
          pattern: '^[0-9]{9}$'
          description: "Must also pass the standard ABA routing-number checksum."
        accountNumber: { type: string, pattern: '^[0-9]{4,17}$' }
        accountType: { type: string, enum: [checking, savings] }
        bankType: { type: string, enum: [bank, credit_union, savings_and_loan, other] }

    Consent:
      type: object
      additionalProperties: false
      required: [termsAccepted, dataProcessingAccepted]
      properties:
        termsAccepted: { type: boolean, enum: [true] }
        dataProcessingAccepted: { type: boolean, enum: [true] }

    CloudinaryUploadTarget:
      type: object
      description: >
        Everything the browser needs to upload one file directly to Cloudinary. apiKey and
        signature are safe to expose — signature is computed server-side with the Cloudinary
        API secret (never sent to the browser) and only authorizes uploading to this exact
        publicId.
      properties:
        uploadUrl: { type: string, format: uri, example: "https://api.cloudinary.com/v1_1/<cloud_name>/image/upload" }
        cloudName: { type: string }
        apiKey: { type: string }
        timestamp: { type: integer }
        signature: { type: string }
        publicId: { type: string }
        type: { type: string, enum: [authenticated] }
        overwrite: { type: boolean, enum: [true] }
        invalidate: { type: boolean, enum: [true] }

    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - INVALID_IDEMPOTENCY_KEY
            - INVALID_IMAGE
            - APPLICATION_NOT_FOUND
            - MALFORMED_JSON
            - CORS_NOT_ALLOWED
            - PAYLOAD_TOO_LARGE
            - RATE_LIMIT_EXCEEDED
            - NOT_FOUND
            - INTERNAL_ERROR
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }